Sokect.IO integration with main proyect
