+++ controllers/chatController.js
@@ -1,48 +1,110 @@
-// const express = require('express');
-// const app = express();
+const catchAsync = require('../utils/catchAsync');
+const AppError = require('../utils/appError');
+// const io = require('./../utils/socketConection');
 
-// const http = require('http').Server(app);
+const Chat = require('../models/chatModel');
+const User = require('../models/userModel');
+const Message = require('../models/messageModel');
+const Interaction = require('../models/InteractionModel');
 
-// const server = require('./../server');
+exports.protectChat = catchAsync(async (req, res, next) => {
+  if (!req.body.message) {
+    return next(new AppError('Por favor, ingrese un mensaje.', 400));
+  }
 
-const ioConnect = (server) => {
-  const io = require('socket.io')(server);
-
-  io.sockets.on('connection', function (socket) {
-    //   socket.on('username', function (username) {
-    //     socket.username = username;
-    //     io.emit('is_online', '🔵 <i>' + socket.username + ' join the chat..</i>');
-    //   });
-
-    // socket.on('disconnect', function (username) {
-    //   io.emit('is_online', '🔴 <i>' + socket.username + ' left the chat..</i>');
-    // });
-
-    socket.on('username', function (username) {
-      console.log('Username event...');
-      socket.username = username;
-      io.emit('is_online', socket.username + ' joined the chat...');
-    });
-
-    socket.on('chat_message', function (message) {
-      io.emit('chat_message', socket.username + ': ' + message);
-    });
-
-    socket.on('disconnect', function (username) {
-      io.emit('is_online', socket.username + ' left the chat...');
-    });
-
-    socket.on('testEvent', (data) => {
-      console.log('Client emited a test event. Data = ', data);
-    });
+  const interaction = await Interaction.findById(req.body.interaction).populate({
+    path: 'offerer',
   });
-};
 
-module.exports = ioConnect;
-// app.get('/', function (req, res) {
-//   res.send('hello world');
-// });
+  if (!interaction) {
+    return next(new AppError('Interación inválida.', 404));
+  }
 
-// const server = http.listen(8080, function () {
-//   console.log('listening on 8080...');
-// });
+  if (interaction.state != 'match') {
+    return next(
+      new AppError(
+        'No tiene permiso de iniciar una conversación con el usuario indicado.',
+        401
+      )
+    );
+  }
+
+  if (req.user.userType === 'applicant') {
+    if (req.user.id != interaction.applicant) {
+      return next(
+        new AppError(
+          'No tiene permiso de iniciar una conversación con el usuario indicado.',
+          401
+        )
+      );
+    }
+
+    req.user2 = { id: interaction.offerer.id };
+  } else if (req.user.userType === 'offerer') {
+    if (req.user.organization) {
+      if (!req.user.organization.equals(interaction.offerer.organization)) {
+        return next(
+          new AppError(
+            'No tiene permiso de iniciar una conversación con el usuario indicado.',
+            401
+          )
+        );
+      }
+    } else if (req.user.id != interaction.offerer.id) {
+      return next(
+        new AppError(
+          'No tiene permiso de iniciar una conversación con el usuario indicado.',
+          401
+        )
+      );
+    }
+
+    req.user2 = { id: interaction.applicant };
+  }
+
+  next();
+});
+
+exports.createChat = catchAsync(async (req, res, next) => {
+  let chat;
+
+  const message = await Message.create({
+    message: req.body.message,
+    sender: req.user.id,
+  });
+
+  const query = { $in: [req.user.id, req.user2.id] };
+  chat = await Chat.findOne({ user1: query, user2: query });
+
+  if (chat) {
+    chat.messages.push(message.id);
+    chat = await chat.save();
+  } else {
+    chat = await Chat.create({
+      messages: [message.id],
+      user1: req.user.id,
+      user2: req.user2.id,
+      isLive: true,
+    });
+
+    let user2 = await User.findById(req.user2.id).select('+chats');
+
+    req.user.chats.push(chat.id);
+    user2.chats.push(chat.id);
+    // req.user.chats ? req.user.chats.push(chat.id) : (req.user.chats = [chat.id]);
+    // user2.chats ? user2.chats.push(chat.id) : (user2.chats = [chat.id]);
+
+    await req.user.save();
+    await user2.save();
+  }
+
+  const io = req.app.get('socketio');
+  io.emit('chat_message', message.message);
+
+  res.status(201).json({
+    status: 'success',
+    data: {
+      chat,
+    },
+  });
+});
